m.htmlsanity: introduce and use new format_siteurl filter.
I needed a way to prepend SITEURL to page, article, tag etc URLs that's
consistent with current (wrong) Pelican way of doing it:

    '/'.join([SITEURL, url])

That of course doesn't work when `url` is an absolute URL, so I'm using
urljoin (but still staying compatible with the above). This is now also
used throughout the template -- I mistakenly assumed that when SITEURL
is set, all page.url etc. properties already have it, but that's
apparently *not* the case, which caused me a bit of headache the past
months.