css: set the hamburger glyph in top navbar from CSS.
Instead of hardcoding it from the markup. This way when the CSS is not
present at all (broken site), the user is not present with confusing
hamburgers on the top.

Important: unless you are using the builtin Pelican theme, you need to
update your markup to remove the hamburger glyph, otherwise you'll have
two of them there.