css: improve code background color.
In the dark theme it's still the same as header background color, but
with 50% transparency. To avoid unnecessary transparency and browser
perfomance issues coming out of that (yes I *do* care about that), a
pre-calculated value combining it with background color is used.

Also, until now, code inside colored notes was looking awful. Now it's
using the 50% transparency and looks good. In order to avoid combinatory
explosion of variables, the transparency *is* used in case of notes, but
since cases like that are much less common, I don't think this is a
problem.

The light theme is not changed at all, except for adding a matching
semi-transparent variable.

Huge thanks to @Squareys for this suggestion.