css: rework button markup.
Default to centered button, because that's what one usually wants. This
also resolves the problem with the following markup not having spacing
after and thus being too sticky to content that immediately follows it.

    <div class="m-text-centered">
      <a class="m-button m-success" href="#">...</a>
    </div>

The new markup is thus this, the .m-text-centered class no longer
needed. Adding .m-fullwidth to the <div> works as expected.

    <div class="m-button m-success">
      <a href="#">...</a>
    </div>

Sorry for potential breakages of your content.