doxygen: search result clarity improvements.
 * Function/macro params and suffix (such as const&&) is included in
   search results.
 * Directories are displayed with / at the end.
 * Parent pages are included in the name.

In order to properly highlight the typed value, a new "suffix length"
field is introduced in the result map, saying how much text there
is after the matching trie entry.

Also, in order to display the right angle quote for subpages, a bunch of
UTF-8-awareness fixes was done across the code.