doxygen: add compound to search data during the actual parse.
So it's prepared for adding extra search stuff that needs to be parsed,
such as search keywords.

I went through the test case and manually reordered by hand to ensure no
functionality was lost while doing this. It was not, quite the opposite
-- groups were not properly marked as deprecated. Now they are. The file
size stays the same, as it should.