doxygen: verify that \param etc. is extracted out of a list as well.
Previously this was asserting because I didn't expect Doxygen to be so
shitty and put these *inside* of some random list without ending that
list first. This also uncovers a lot of other nasty Doxygen bugs, so I'm
just creating a very restricted repro case that doesn't trigger them.

I lost all hope with Doxygen. Kill it with fire and start over.