doxygen: put search icon size directly into the markup.
Fixes the problem with a giant magnifying glass when no style is present
and also should fix the random layout bugs with top navbar being twice
as high (which is probably caused by the browser not calculating the
floating etc. properly due to such a huge change after the style is
loaded).