m.htmlsanity: clarify the use of expand_links a bit.
I wanted to remove it at first, but then realized a more proper solution
would be to not require the user to add random fields to
FORMATTED_FIELDS. In the ideal case the template *could* do just

    {{ field|render_rst|expand_links(article }}

but at the moment Pelican strips all reST formatting from the
non-formatted fields, which kinda prevents this. So keeping the filter
for later when we have this fixed.