doxygen: hide the search by going back only when it makes sense.
If we landed directly on an URL with #search, going back would mean the
browser is redirected back to the referrer page, the Speed Dial or
whatever. Do not want. In this case we simply replace the history state
without the #search hash -- there definitely wasn't any other hash we
could return to.