Moved doxygen/dox2html5.py to documentation/doxygen.py.
And move the theme and test directory as well. Making room for Python
and other language support. Note this is a breaking change --- from now
on, you need to call the script as

    documentation/doxygen.py

instead of

    doxygen/dox2html5.py

Besides that, the location of default templates changed from
doxygen/templates/ to documentation/templates/doxygen/ --- if you have
your own modifications to these, you may need to move your files.