Moved plugins from pelican-plugins/ to plugins/.
Not just for Pelican for a long time already. Note this is a breaking
change --- if you're using the m.css Pelican theme, you need to update
your PLUGIN_PATHS to reference the new directory, for example:

    PLUGIN_PATHS = ['m.css/plugins/']

instead of

    PLUGIN_PATHS = ['m.css/pelican-plugins/']

If you're using the Doxygen theme or just the CSS files, nothing changes
for you.