documentation/python: rename brief to summary.
Another doxygen-ism. It made some sense for API docs, but having "brief"
and "description" for pages doesn't make any sense at all. Summary and
content makes much more sense for pages and since it doesn't make sense
to have two types of naming, let's rename it for everything else as
well. This is now also in-line with Pelican article metadata and XMLDoc
for C# docs, yay.