documentation/python: properly handle recursively imported modules.
Basically converting the crawl to be breadth-first instead of
depth-first and maintaining a set of crawled modules to ensure they're
never visited again.

This could happen with classes as well, but it's rarer -- one would need
to add it to __all__ twice (or once and then it gets discovered
implicitly) etc. For now I'm simply asserting, will revisit in case it
proves to be an actual problem.