documentation/python: less responsibility on docstring-processing plugins.
Until now, the plugin was expected to fill both summary and content for
the symbol it's ccalled on. THat made plugin implementation harder than it
should be, causing a lot of code duplication and explosion of testing
effort.

Now the script simply sills in summary / content if it isn't there after
all processing and that's it. No asserts anymore.