documentation/doxygen: support \since as labels in entry lists.
Makes everything much easier to discover. Since I'm by far no longer the
only user, I have a fear of breaking things, so this is completely
opt-in.

Also recognizes a combination of \since and \deprecated as "a version in
which given API got deprecated", expanding the red deprecated label with
a custom text.