documentation/doxygen: allow \xrefitem with numbers in the middle.
In Magnum I used to have \xrefitem aliases such as \requires_gles20 so
the current regex worked, but when adding \requires_gles20_only it
didn't work anymore. Such a stupid oversight, sorry to whoever else
tripped up on that as well.

Not updating any test case because regexes are write only anyway.