package/ci: drop Python 3.5 testing, add 3.8 to 3.10.
Python 3.5 was EOL'd in September 2020 and support for it is being
removed from major packages like pybind. It was also the only version
on which neither the Doxygen nor Python doc generator worked due to lack
of typing annotations, and since those two are the major use cases, I
don't think there's many users relying still on 3.5.

Python 3.6 was EOL'd in December 2021 so *technically* I could drop it
as well, however Ubuntu 18.04 LTS still has it as the default version so
it's good to keep it. I don't really need to use any 3.7-specific
features yet either, so this isn't a problem.