m.plots: update to work with matplotlib 3.5 as well.
The major change in 3.5 is that the attributes have a different order
and styles have more whitespace and less semicolons. To avoid explosion
of find/replace patterns, the search strings were converted to regexps
-- which on the other hand allowed me to drop some duplicates, and
discard the font name replacement altogether.

On the CI I'm now using the latest version on Python 3.8+, 3.7 stays on
3.4 to avoid regressions on that versions, and 3.6 on 3.3 which is also
a bit different.