package/ci: I HATE WEB DEVELOPMENT
All packages getting changed all the time with stupid differences. IT IS
NOT POSSIBLE TO JUST HAVE A PROJECT WORKING FOR SIX MONTHS WITHOUT A
CONSTANT NEED FOR ADAPTING TO CHANGES IN DEPENDECIES FFS!!!

I miss developing against Qt, where the APIs I used in 2007 were still
there in 2014. ARGH