m.code,doxygen: test inline ANSI code highlighting.
Because it got broken with Pygments newer than the ancient 2.6.1 fork I
have (it now unconditionally adds a wrapper <div>), and unfortunately I
can't see how to not do that in 2.14 without delegating into private
APIs.

So just to have that problematic part reproduced in tests, once I get to
that. Unfortunately latest Pelican doesn't work with older Pygments
anymore for some reason (or, well, disallows it), so I can't generate
the website until I have this fixed properly. Ugh.