m.plots: replace hashes in test files before comparing.
Similarly to what I did for doxygen, the damn thing now has hashes
differing even between machines. How the hell is one supposed to make
reproducible builds with that, I don't know.

The 3.4 variant got removed, because it differs from 3.2 only in those
hashes, so it isn't needed anymore. Now I can finally unpin it from 3.7
on the CI and stop being afraid that it changes again with 3.9.3.