pelican-theme: don't attempt to query a Pelican version in the test.
Because I rely on ancient pygments (gotta fix that too), I cannot use
Pelican 4.9.0+ as it requires a newer version. And on Python 3.12
pelican.__version__ with Pelican 4.8.0 says "unknown" (probably due to
not being ported to the new packaging thingy or whatever), breaking the
version parser in the test.

I'm querying the version just to skip a test that relies on a PR that I
opened in 2017 and didn't have time to get back to. I likely won't have
for another decade, so can just skip this always.