documentation/doxygen: properly escape HTML <>s in a test.
This was making Doxygen 1.10 + 1.11 fail parsing the link title,
substituting for an empty text. Funnily enough 1.12 doesn't suffer from
that anymore, which is what made me think that was a Doxygen bug and not
mine. Well it still *is* a Doxygen bug, but at least now I can have
consistent output everywhere.