m.code: adapt to pygments 2.10 to 2.18.
The new versions properly qualify whitespace as such instead of as text,
and do so gradually, which is responsible for most changes in the
output. Besides that, until 2.14 there were certain cases of empty
<span> in the output, which 2.14 fixes, so I add a postprocessing step
to clean those up to not need to further branch the tests.

The CI now tests various pygments versions, with only the last job
taking the latest version.