documentation/python: adopt PEP585 naming for pybind11 typing hints.
This is what pybind11 2.12+ does, but as I have to parse the docstrings
and prepend the `typing.` namespace, I can retrospectively make this
change for older pybind11 versions as well. In particular, the following
is changed in type annotations of generated documentation, independently
of pybind11 version:

 - typing.Tuple is now tuple
 - typing.List is now list
 - typing.Dict is now dict
 - typing.Set is now set

If the python standard library inventory file is used, the type
annotations in signatures will now link to the builtin types instead of
the typing module.

Note that if your documentation uses external docstrings and matches
overloads by the signature, you may need to adapt the rename to make
them match again.