documentation/doxygen: this warning isn't related to any file.
It may get printed when preprocessing the state from all files (building
parent/child hierarchies, etc), listing a random file for that is
extremely confusing.

Well, the whole message is extremely confusing, so let's redo it,
suggest a fix and also have a test for it, instead of excluding it from
coverage.