documentation: verify HTML escaping for both the doxygen and python theme.
It's a mess and needs to be cleaned up, especially given that I now need
to have non-HTML output from the Python generator, so it's good to have
it actually regression checked.

This actually revealed one superfluous unescape in the Python theme
(which was likely a leftover from when that code was copypasted from the
Doxygen theme), and a Doxygen bug where it doesn't unescape image alt
text. Then m.css doesn't correctly escape the alt text either, which
makes the bug cancel out and everything is alright.

Oh and there's also a bug where if special HTML chars are used in
filenames, it breaks the Doxygen XML output. I should probably make a PR
for those two.