documentation/python: expose also plain relative types to the template.
Not just type links, but also plain types with redundant name prefixes
removed, for use in contexts where a link doesn't make sense, such as
tooltips. They're HTML escaped like the other variants, but right now not
used in the template at all.

The only breaking change is that what was `param.default` / `data.value`
is now in `param.default_link` / `data.value_link`. Those are different
from `param.default` / `data.value` only if the type is a (known) enum,
in which case the link is to the enum value, other than that these two
are the same.