documentation/python: when signature fails, leave it up to the template.
Passing an ellipsis to the template made sense when there was just a HTML
output, but for generated stubs that's not a valid syntax. So output a
single parameter with name being None instead, and let the template
decide what to do there.

In case of pybind, the code was inconsistently using the Unicode
ellipsis. Now it returns None too, and thus the template handles both the
same.