documentation/doxygen: fix bad interaction with VERBATIM_HEADERS=NO.
Overriding include file in d3e6203dad8a5c6a87554f504c924ef36aa80ba6
unfortunately broke with configs that have VERBATIM_HEADERS disabled
(which is a sane thing to do, as those don't get used in the output at
all). However in that case the <includes> don't contain refid, which
means it's not really posible to link to anywhere.

Give up in that case and use just the original un-overriden name,
printing a warning. I *could* support some halfway solution, where the
name is overriden but it links to some other place, but that's yet
another code path to verify. Just don't.