documentation/python: improve handling of broken pybind11 signatures.
If certain overloads from an overload set failed to parse and search was
enabled, the code died on an assert because the search data population
didn't expect overloads with broken signatures, and with just one
function it doesn't use the arguments at all, so this was quite a corner
case.

When testing the above, I ran into a case where the builtin math.log()
docstring was attempted to be parsed as a pybind11 docstring which it
isn't, so another graceful handling is now in place for this to not
happen again.

Finally, the `...` ellipsis is used for unrepresentable / too long
default values, and that's valid Python code. But it also got used for
the broken signatures (which isn't valid Python), so now I'm using the
Unicode `…` in that case instead to distinguish the two.